web/requirements.txt
flask>=3.0.0
google-genai>=1.25.0
jinja2>=3.1.6
ollama>=0.5.1
python-dotenv>=1.1.1
